Updated openstack/openstack Project: openstack-infra/project-config  25d41596b0fb248ac5309ff1505a494b622cf24c
Rename puppet project 'puppet-openstack_health'

This commit renames the module to be 'puppet-openstack_health'.

As Spencer Krum (nibalizer) pointed out at https://review.openstack.org/#/c/227849/
this project is not following the naming conventions for puppet modules.

This has a major impact on install_modules.sh, which will not be able to clone and
install this module appropriately.

As part of this, we will need to schedule a maintenance window to take down
gerrit, zuul, etc. and follow the steps enumerated at:
http://docs.openstack.org/infra/system-config/gerrit.html#renaming-a-project

Change-Id: I10c0b61268179088496fe3c381ebd17935cc9e69
Co-Authored-By: Caio Carrara <ccarrara@thoughtworks.com>