Updated openstack/openstack Project: openstack/requirements  907ecec9226887af05e7d8a1985a4e8ae682d199
Bump minimum os-testr version to 0.4.1

The 0.4.0 release added a new whitelist-file option which lets you
specify a list of regexes to run the tests for. The 0.4.1 fixes a
problem with the blacklist-file option as well.

Nova will leverage one of these new feature to start adding py34
tests incrementally, so we need to bump the min. required version.

Change-Id: Ib5e682d0380cf7bc5e288a1e4d125379b452fa92