Updated openstack/openstack Project: openstack/requirements  14f478c7d8bacafd4111a615ce22dfe85a63ff6f
Add simple doc publisher target

Add a simple doc publisher target that just includes the existing
README.

Some of the infra jobs that fail requirements try to point you to the
requirements documentation; currently this is given as a raw RST link
to the web-view of the git tree.  It would be nicer to publish this as
formatted HTML like all the other docs.

Change-Id: I56b60a950c64880dd551d11e1997cd5ecbae283c