Updated openstack/openstack Project: openstack/api-site  707228c88d02592d91e1c788e8eac6015a5cf181
Add a description of an optional parameter "id" of "Create region"

Add a description of an optional request parameter "id" for
"POST /v3/regions".

Add a note of "id" because it has to be unique to the Openstack
deployment.

For details, please see http://developer.openstack.org/api-ref-identity-v3.html#createRegion.

Change-Id: I8e733d513cd42a31250df15c954ad80d1c168ab6
Closes-Bug: #1501744