Updated openstack/openstack Project: openstack-infra/shade  3e4bcd0a6cd8e0736bd6defeb311f99e69c9a8ce
Return IPv6 address for interface_ip on request

If the local host looks like it can route IPv6 and we have an IPv6
address, and the user has not indicated that they must use IPv4 for
some reason, return it to them in the interface_ip field.  Also, always
return something in interface_ip if possible.

Change-Id: I3544280cab7abfa6b4154244f4c4588bc65c7347
Depends-On: Ib0c4c4e8b3b7b4bcee5fa3414719969274929b9a
Co-Authored-By: James E. Blair <jeblair@linux.vnet.ibm.com>