Updated openstack/openstack Project: openstack/cinder  9b2ed867a5cd8752d4c5086b79308a9a3c95213f
Fix enable/disable_replication raise InvalidVolume

In both enable_replication and disable_replication,
it tests to see the replication_state is in the correct
state.  When it's not it was raising InvalidVolume()
InvalidVolume requires a reason text message.

This patch adds the missing reason=msg when creating
the InvalidVolume exception

Change-Id: I66a3e4dc4359457e9669c5ddaa4b0529645334f9
Closes-Bug: 1501942