Updated openstack/openstack Project: openstack/neutron  a2cfafbb0c7c77b8e38b075d3ccb372c0383f398
Do not log an error when deleting a linuxbridge does not exist

When you're racing to teardown and relying on RPC, it should not be
considered an error that a bridge does not exist when you're trying to
delete it anyway. This is better served as debug level information.

For context, this shows up over 1000 times in successful
gate-tempest-dsvm-neutron-linuxbridge job runs in a week so it's
probably safe to say it's not contributing to test failures.

Closes-Bug: #1501779

Change-Id: I8e9b27e010d033291497f4e660a97bbae92e75b1