Updated openstack/openstack Project: openstack/neutron  82866276b0cf7a10cbb9f5e92a12680b096a3770
Deprecate branchless migration chains from neutron-db-manage

Support for the scheme puts additional burden on neutron-db-manage
maintainers.  The split branches were introduced in all 'official'
subprojects in Liberty, so it's fine to deprecate it in Mitaka, and drop
in Noodle.

Since there is no user actionable item to do for users, using
debtcollector that will only issue DeprecationWarning if they are
enabled. The assumption is that developers run their tests with that
enabled (as the BaseTestCase ensures) and handle warnings.

Related-Bug: #1501380
Change-Id: Ie4ddd29d8c51be74a112864aae3d16fb5e52c0fa