Updated openstack/openstack Project: openstack/nova  2bb567a5ab0102580d7cbd7044dc63ffe64944e7
VMware: raise NotImplementedError for live migration methods

The VMware VC driver currently does not support live migrations.
This was supported by the ESX driver but that was removed in the
Juno cycle.

In the mean time the driver will raise a NotImplementedError for
live migrations until we fix this.

Related-Bug: #1192192

Change-Id: I20cdbb5892d2e45b0b8f9a658793ab20dbf48fa1