Updated openstack/openstack Project: openstack-infra/project-config  e4e6a87afff607da06b53e1161e538c9ea5956f9
Add a nodepool + shade from src integration job

We want to do integration testing of nodepool and shade from source
installs. First step in doing that is getting devstack-gate to checkout
the correct versions of both shade and nodepool on disk. Do this in a
new job (preserving the old test nodepool against released shade job) so
that we test future development and how people will deploy nodepool in
production.

Change-Id: If323a7ab577c7942f4f98a42bbbb291c86d5b6f7