Updated openstack/openstack Project: openstack/heat-translator  c3aa4c22e9dbd4969cdd884c2f9626c3f28ca0f4
Fix custom import issues when using shell environments

Translator is enabled to exectue within other shell environments by
installing as a library (in addition to command line). As such, a
separate package "translator" is installed, and pulls in everything
under the translator/ directory.

Originally the design for custom classes was to allow users to specify
their own locations for custom classes that they define outside of the
translator code. In order for this to work in the shell environments,
the user would need to install these classes via a setup.py script.
For now this is too heavy of a requirement to place on the user, so
the design for this fix is that the user will need to manually copy
over their custom classes to the translator/custom directory.

We will revist the option to install custom classes via a user defined
setup.py at a later date.

Change-Id: Ib132d9fe768a954a226724929c074034cf898da3