Updated openstack/openstack Project: openstack/nova  251e09ab69e5dd1ba2c917175bb408c708843f6e
libvirt: check if ImageMeta.disk_format is set before accessing it

Commit 4f43bbafa4bfc2ab149d5db4eb55320c1602658b changed
blockinfo.get_root_info to access the disk_format attribute on the
passed in image_meta (ImageMeta object) like an object, but if that
attribute is not set you'll get an error.

image_meta can be an empty object in the case of evacuate where
image_ref is passed as None to the compute manager's rebuild_instance
method in the case of evacuate, so image_meta defaults to {}. So when
the ImageMeta object is created in LibvirtDriver.spawn, there is nothing
in it.

We have to check if the attribute is set before accessing it using
obj_attr_is_set because the ImageMeta class does not include the
NovaObjectDictCompat mixin, so we can't use
image_meta.get('disk_format').

Closes-Bug: #1501831

Change-Id: I95e543415b881603bb997ae0e159017c03e58a53