Updated openstack/openstack Project: openstack-dev/devstack  c4564dfbe60a57069679ff29d488543dc34b0c6f
Install python db client if a db backend is configured

If `DATABASE_TYPE` is configured in `local.conf`, the database backend
is currently configured with `initialize_database_backends` even if no
database backend is enabled.

On a multi-nodes Devstack environment, such as devstack-vagrant, the
compute node currently fails because it does not have PyMysql. This
compute node has no database backend enabled, but has to connect to
the database on another node.

We should install the python client if DATABASE_TYPE is set, even
if no database backend is enabled.

Closes-Bug: 1501001
Change-Id: Iffd5f7243a0dfdbe56cf6b9a87b96ed7678c81dd