Updated openstack/openstack Project: openstack/openstackdocstheme  cd4ea5aed8571a3862090fd0315eee8db5e293e7
Add support for Firstapp

This patch enables Firstapp documentation to use the log-a-bug
feature of openstackdocstheme, which was previously only working
with OpenStack manuals.

The relevant code is in script_footer.html and is extensively
documented.

In order to allow testing, I added part of the api-site tree
to the openstackdocstheme repo.

Change-Id: Ic5e8dd6733f26a5be00b242a3d3203debb19e8c2