Updated openstack/openstack Project: openstack/openstack-manuals  bfc22430e3db087028585a7cddf21bb55deac2ac
Provide specific path of the 'local_settings' file

The path of the 'local_settings' file is not right in
"http://docs.openstack.org/kilo/config-reference/content/configure-dashboard.html".
The path of the 'local_settings' file is different under different
Linux versions. This patch give the specific path of the "local_settings".o

Change-Id: I1dcea4e58d20f7c8cc2a1c397d75ac3e6a667486
Closes-bug: #1502609