Updated openstack/openstack Project: openstack/oslo.utils  9ec6cccb77a1e8ec5ba0b0947b3ba5e16502a2db
Make forever_retry_uncaught_exceptions handle its own failures

When an exception occurs inside 'except' clause, it is not handled.
As a result, forever_retry_uncaught_exceptions fails with exception,
while by definition it should not.

For instance, oslo.messaging's RPC server relies on that
function to process any exception. When forever_retry_... fails
to do so, the server thread dies. An example could be
found in referenced bug.

Change-Id: I415a0f49b25b80a264f0bc951f4b926d57a9c9a8
Closes-Bug: #1502092