Updated openstack/openstack Project: openstack/openstack-manuals  ec25954df1ee498bf8990f935b0f835bc4d92355
Fix NTP verify appearing before install

Due to the way the toctree was working on this page, the
steps to verify the NTP installation:
a) appeared prominently on the initial page, leading to
people wanting to try them before installing the service
b) were not available in a logical order when using the
previous/next buttons

This patch adds a new file for the verify steps to fix a)
and b) and includes it from the toctree. It's a solution,
but may not be the best solution.

Change-Id: I14b977147e70bf29ea42e40bb337c89426e6b8d5
Closes-Bug: 1501704