Updated openstack/openstack Project: openstack/python-novaclient  b7e647427df9f883b21ca741da830051e962411b
Test that microversions are not skipped

When raising API_MAX_VERSION to support a new Nova API microversion
there should be support added to some method to account for the change.
A test has been added to catch if support has not been added.  In the
event that no methods need to change there is an exclusion mechanism in
the test so the version and a note can be added there.

Change-Id: Iebc1be2557e1b6eec327b703568805578e564172