Updated openstack/openstack Project: openstack/nova  cd8e1f175e4f07984071e850094a8f61fca83048
Add checkpoint logging when building an instance in compute manager

When the large ops job fails because instances timeout during the build,
it's hard to know what is going on for a given instance in the compute
manager's _do_build_and_run_instance flow since a lot of the trace
logging that happens is in the virt drivers.

The large ops job uses the fake virt driver which doesn't do much
logging at all. Rather than add logging to the fake virt driver, it'd be
useful to add some logging to the compute manager when we are doing
sub-tasks like building networks, building bdms, and spawning the instance
on the hypervisor.

Per: https://wiki.openstack.org/wiki/LoggingStandards

The 'Starting instance' log message is changed to debug level and the
completed task logging unit of work messages with time spent are done at
info level.

Related-Bug: #1500615

Change-Id: I6b9e7fc7b3a14ccf8b1531fa08ce8d732e23775a