Updated openstack/openstack Project: openstack-infra/nodepool  2e05f1850f811efbd1a802c06f8de3cdabd7cc3b
Restore ability to run nodepoold with fakes

With OSC and shade patches, we lost the ability to run nodepoold
in the foreground with fakes.  This restores that ability.

The shade integration unit tests are updated to use the string
'real' rather than 'fake' in config files, as they are trying to
avoid actually using the nodepool fakes, and the use of the string
'fake' is what triggers their use in many cases.

Change-Id: Ia5d3c3d5462bc03edafcc1567d1bab299ea5d40f