Updated openstack/openstack Project: openstack/nova  5732a5cd11b190845e7b9fe0344cadec16e4f891
Change parameter name in utility function

Change b7e9a64416ff239a4c1b8501f398796b02c46ce7 introduced a
filter_deleted flag into the metadata_to_dict utility function.
Counterintuitively, when this flag is False, deleted items are excluded
from the returned dict, and when the flag is True, deleted items are
included.  (The default setting is False, which preserves the behavior
of this function before the flag was added.)  This poor choice of names
doesn't sound like a big deal, but there are some functions calling
metadata_to_dict that are using the flag improperly (that is, they are
setting filter_deleted=True in the expectation that deleted items won't
be included in the output). This patch changes the name of the parameter
to the more perspicuous 'include_deleted', sets the default value to
False to preserve the original behavior of the function, and adds some
tests.  (I'll correct the functions that are using the flag incorrectly
in a dependent patch.)

Change-Id: I40737ebb19fb50e3f671c996a2d096996835cc68
Closes-Bug: #1496394