Updated openstack/openstack Project: openstack-infra/infra-specs  f4bb6335ca5ecacd2cf2f232f5c29392928c142f
Host an initial CI monitoring dashboard service

Project communities requiring third party CI need to be able to quickly
see the status of a CI system.

As a service to our developers, this spec proposes a quick solution to
host existing CI Watch codebase as a third-party CI monitoring
dashboard.

Change-Id: I4f77a0af6280964d70632ec4978f353085f2d56b