Updated openstack/openstack Project: openstack-infra/nodepool  b7cc4b894b843897cc18c8a40954c2dea95994a5
Move image location out of config location

In the devstack plugin don't write the images to the config dir for
nodepool. This prevents us from copying the built qcow2 to the log
server (saves disk space and speeds up the jobs). This also provides
logical separation from the inputs (config) to nodepool and the outputs
(images) of nodepool.

Change-Id: I066ccda8a564bf9ea144a3dda9f6d506bb130a61