Updated openstack/openstack Project: openstack/python-saharaclient  975387cdf5a7464d44298d3abeba0601d7a543a3
print usage when no argument is specified for python3

When running just sahara' under python3, we will get the error:
ERROR: b"'Namespace' object has no attribute 'func'"

This is because map() is used to decode sys.argv, but under Python3
it returns a map object which is an iterable. Some code later tries
to use this in a boolean context and it will always return True,
even if it's empty.

Change-Id: I508b093e30814645b2bd4c32cf469204bb696207
Closes-Bug: #1295356