Updated openstack/openstack Project: openstack/requirements  07fc4dd0773a9e92c44eff99f1d58f33d4a10859
Add weakmethod backport library (for 2.7 and 2.6 only)

Python 3.4 has a weakref.WeakMethod class which knows how
to correctly reference instance methods, to enable weakrefs
to instance methods in 2.7 and 2.6 correctly this backport
is needed (or an equivalent).

- Is the library actively maintained?

Seems to be, released in last year.

- Is the library good code?

Just a backport from 3.4 so little/no expected variation over the
code in python 3.4.

- Is the library python 3 compatible?

Yes

- Is the library license compatible?

PSF license, so yes.

- Is the library already packaged in the distros we
  target (Ubuntu latest / Fedora latest)?

Does not seem to be on ubuntu or fedora yet.

- Is the function of this library already covered by
  other libraries in global-requirements.txt?

Does not seem to be, `six` does not include this (which would
be the closest library that might).

Change-Id: I3ac7ba5eaa1de3cee5739287703dd0b3565a7484