Updated openstack/openstack Project: openstack/openstack-manuals  295669fc988ba15c63582ca73d3f2ef583c57dc6
Give specific path of the 'openstack-dashboard.conf' file

The path of the 'openstack-dashboard.conf' file is not right in
"http://docs.openstack.org/kilo/config-reference/content/configure-dashboard.html".
The path of the 'openstack-dashboard.conf' file is different under different
Linux versions. This patch give the specific path of the "openstack-dashboard.conf".

Change-Id: Idde862aacaf114e30befc4c8edb9b0f006ad41b9
Closes-bug: #1502927