Updated openstack/openstack Project: openstack/heat  26be3cde576f4ad0ab9cf4b1503447482e3112dd
Align the docs intro text with the old wiki page

This landing page is now aimed at users, operators and developers
(instead of just developers).

This change rewrites the intro section to also cover points made
in the old wiki landing page [1] which needs to be deprecated with
a link to this page instead.

[1] https://wiki.openstack.org/wiki/Heat

Change-Id: I452807a2d9a8821c4a81d63eb8e824e04b344ef5