Updated openstack/openstack Project: openstack/pycadf  7bd04962783700c91593a7b1bf6ba2e1108b4671
Change ceilometer endpoint name from telemetry to metering

The pycadf ceilometer api_audit_map.conf file currently refers
to ceilometer service_endpoint as telemetry and not as metering.
This is not aligned with the rest of the ceilometer community code.
The service_endpoint was named as telemetry initially assuming
the entire ceilometer code will switch to using it but that has not
happened yet. Thus, it's better to stick to metering for now.

Change-Id: I053177ef504989c4b9b367d671d47b00104600ab
Closes-Bug: 1501287