Updated openstack/openstack Project: openstack/tempest  c920ba8187a91ea8715600a3091c1f0604065847
Deprecate credential config options

This commit deprecates the original configuration options for
providing credentials. This mechanism of providing tempest with
credentials to run with has a lot of drawbacks and has been
superseded by using an accounts.yaml file. It has been planned
for deprecation once we had a periodic/experimental CI job setup
to run with test accounts enabled. Now that this has been done
we can start the deprecation cycle for the old options.

Change-Id: Ia5132c5cb32355d6f26b8acdd92a0e55a2c19f41