Updated openstack/openstack Project: openstack/ironic  edda2f26246b0f93a2f318462b7ee142a2c59ca8
Add notes to functions which are in ironic-lib

We are planning to use functions which are now in ironic-lib. Add notes
to functions here which correspond to functions in ironic-lib. To help
developers make sure to update both functions until the functions have
been deleted from ironic and we start using ironic-lib instead.

Change-Id: I943f8170c0dbc62dc8325d48633a939a2726eb4f