Updated openstack/openstack Project: openstack/python-openstackclient  e9269987005d1ac496f0bd0e9345375ca8ea689e
Image fix bug with --volume

Currently after calling the cinderclient to create an image from a volume,
it also then tries to create another image. This fails as the keyword
volume is unexpected. This add checks so the image is not created in
this case. Allowing --volume to not throw an error when it has worked.

Change-Id: I67e650eb0b8c331d86515e3e326c39a5d6dad5e1
Closes-Bug: 1497221