Updated openstack/openstack Project: openstack-infra/project-config  059f7e771382c56ab85fdbc25536f0ad8d84a80d
remove extra ceilometer devstack plugin declaration

The ceilometer devstack plugin was getting declared in project-config,
as well as by the ceilometer grenade plugin. This caused the
ceilometer project to be attempted to be run more than once. We should
probably protect against that, but we should also not declare it
twice. Especially as ceilometer is the first project to get all this
going, and it's configuration will likely be copied into others.

Change-Id: I9f083731b2ffe4a98cf7bfd5512a4aa3697af2fa