Updated openstack/openstack Project: openstack/nova  c76733f6d92aac59f26e028d49d4bd57aab870c1
Allow automatic determination of compute_rpc version by service

This adds a upgrade_levels.compute='auto' option to let the compute RPC
module automatically determine what version to use based on the service
versions in the deployment. By default, we still use the newest version
if not set in the config, a specific version if asked, and only do this
automatic behavior if 'auto' is configured.

If a minimum service version is found that is newer than what we support,
initialization of the compute rpcapi is aborted as this indicates an
ancient node that has been resurrected on an older version of the code
than can be supported by the rest of the nodes.

Also note that this adds the compute RPC version 4.0 element to the
service version history, which was correct for Kilo and will enable
us to do the right thing. Normally, we should not edit this history,
but this is just filling in the value that would have been set when
it was added.

DocImpact: New valid value for upgrade_levels.compute: 'auto'

Related to blueprint service-version-behavior

Change-Id: I679ebedb739a487e2e4e811c4e89b9c07c8b99bc