Updated openstack/openstack Project: openstack/oslo.versionedobjects  e8a758907b4b427079071487d5574442bc7bbb98
Fix obj_to_primitive() when used with a manifest

Previously, we required any child object version bump to come with a
bump to any objects that included it as a parent. That allowed us to use
the version of the object at the top of a tree to determine whether we
should do a backport as part of a to_primitive operation. With manifests,
that's no longer the case. So, if we have been requested a version other
than the current one, OR if we have a manifest, we need to run the make-
compatible step. We will end up doing this all the tiem now, but it should
be very quick if nothing needs backporting.

Change-Id: I27b8a13dadc11b78c890b54eaeb9eb728125977a