Updated openstack/openstack Project: openstack/swift  84687ee3249bd4c358bd1528b0ce1aec78a648f1
Change POST container semantics

To achieve last modified header on GETorHEAD container
we should allow POST container to overwrite put-timestamp.
That is because following patches will change the the put-timestamp
value semantics in both container db and account db as "container
last modified".

If we achieved this and followings, we will be able to retrieve the
container timestamp which is suggested when the container metadata
modified.

Example:
- Create a container.
- Change the ACL with POST container. (e.g. x-container-read)
- After that, we can know the container was changed by comparing
  between the timestamp from container creation and the last modified
  generated from put-timestamp.

Change-Id: I1a545fcd1896798dfa7cb5e5af97c78f5d7d7e4d