Updated openstack/openstack Project: openstack-infra/shade  d8a70eed84294e9f9d028ed8b36567379933a4ea
Add a private method for nodepool server vars

A previous patch split the work done by get_openstack_vars into
costly and less-costly operations.  Nodepool only needs the
less-costly ones.  We will put that information in the shade API
in some form, but it has not been decided what form yet.  In the
mean time, allow Nodepool to easily call the method to perform
the less-costly operations for now.

Change-Id: I99bec7e4a709ef6b7623414dad5f1453070380f8