Updated openstack/openstack Project: openstack/nova  7907e685a88aca93dd136c3d23a5d9a954d47a26
Only create volumes with instance.az if cinder.cross_az_attach is False

Commit 6060888b58db42eea826939852838f9e1c204d2c changed the boot from
volume flow where source != 'volume' (where Nova creates the volume and
then attaches it to the instance during the server create operation)
such that the instance.availability_zone is always passed to the volume
create API.  That was done because if the cinder.cross_az_attach config
option is False then Nova requires the instance and volume AZs to be the
same when attaching the volume.

However, telling Cinder to create a volume based on the instance AZ can
fail if that AZ doesn't exist in Cinder.

Cinder commit b85d2812a8256ff82934d150dbc4909e041d8b31 works around this
by allowing Cinder to be configured to basically not fail if the
requested AZ does not exist. This is not a great solution though if you
have cinder.cross_az_attach=False in Nova because Cinder ignored your
request.  It's also terrible UX for the end user because they get a
NoValidHost when boot from volume fails.  There is a TODO in this change
to address that later (because exactly how cross-service AZs should work
is still being discussed).

This change addresses the gap introduced by always telling Cinder to
create the volume with the instance AZ.  We should only do that when
cinder.cross_az_attach is False, because that's the strict case where
Nova expects them to match.  If cross_az_attach is True, Nova doesn't
care what AZ the volume is in so we shouldn't try and force
Cinder to accommodate the Nova AZ when creating the volume.

Closes-Bug: #1496235

Change-Id: I3d60c084fd71162a0d4835394f8e819723c1b8a7