Updated openstack/openstack Project: openstack/nova  e7a27f6f82237c316ae5da533840e5f86adf4463
Make service create/update fail if version is too old

This makes the Service.create() and Service.update() code fail if we
determine that we are older than the minimum version in the rest of the
deployment.

This would protect us against a scenario where we had an offlined host
during an upgrade, which became reachable later. The rest of the deployment
would have moved on, sending newer versions of RPC messages than we can
send or receive.

Related to blueprint service-version-behavior

Change-Id: I1f488068486ad6c76453fcf22b93eb2f00e7a7a5