Updated openstack/openstack Project: openstack/nova  f2e2a5891d5f5ff9346e6dc8e4dd0e994485245c
Revert "[libvirt] Move cleanup of imported files to imagebackend"

This reverts commit 9ba70756de326ffaa8be43acfde12cad04ed0af2

The change introduced an UnboundLocalError if we fail to
create the config_drive_image variable. Also, the original
change didn't have any unit tests and came late in the
liberty release so I don't really want to mess with fixing
this given we need the fix in liberty-rc2.

Change-Id: Ia7b70aa139b67cf58b5c0f9fbcd2a4deb465914e
Closes-Bug: #1502961