Updated openstack/openstack Project: openstack/ironic  84d370226d8750cf824e19c93cee9d1a68b99acd
Rename 'message' attribute to '_msg_fmt' in IronicException

The 'message' attribute for exceptions has been deprecated since Python
2.6 and does not exist in Python 3. The IronicException 'message'
attribute will still exist (but should not be used) in Python 2.x as the
BaseException class provides it, but in Python 3.x it will no longer
exist.

The preferred way to get a message from an exception is
six.text_type(exc) as this provides the unicode version of the
exception.

This change renames the 'message' attribute to '_msg_fmt' to indicate to
users that they should not use it directly.

For any third-party users of the IronicException class, log a warning
that usage of the 'message' attribute is deprecated, but allow them to
keep working.

This is similar, but different, to work done in Nova:
  commit 70569ae344bceb3794713abc5cb2c82e9671c37d
  Author: Mark McLoughlin <markmc@redhat.com>
  Date:   Mon Jul 1 11:31:06 2013 +0100
  Subject: Avoid shadowing Exception 'message' attribute

Change-Id: Id080ea42d7ae6337ce23c1411cf4e8a903223d3d