Updated openstack/openstack Project: openstack/oslo.messaging  d9bb2f4b94dce727427e4fe57a52214e606c1fb0
ConfFixture should work even when zmq/redis is not present

Previous driver and other locations in the code we used
zmq_async.import_zmq() and importutils.try_import() to
not fail fast when pyzmq and redis python packages are
not installed. If we don't do this, we are going to break
Nova/Cinder/Glance etc.

Closes-Bug: #1503499
Change-Id: I1529241b5e1d902b37d6b610646a5a46a18f13b0