Updated openstack/openstack Project: openstack/openstack-manuals  2869753de1b61b4e30667f96c9b41b4cc300c28d
Removed full path from conf file reference

In the original text, the full path was shown, which was incorrect for some
distros. Patch 231028 gave distro-specific paths, but we do not give
distro-specific information in this guide, and that patch has been reverted.
This patch fixes the original text to remove the full path, and replace it
with the name of the config file only.

Change-Id: I65726323ca15e53dc8d25a4e93627dfc6005cb46
Closes-bug: #1502927